refactor: :truck: 폴더명 변경 filter -> beer-filter
