+++ src/components/BeerListFilterAndSorter/BeerListFilterAndSorter.tsx
@@ -3,8 +3,10 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 
 import Icon from '../commons/Icon';
 import { HEADER_HEIGHT } from '../Header/Header';
-import BeerListFilterChipList, { BeerListFilterChipType } from '../filter/BeerListFilterChipList';
-import BeerListFilterBottomSheet from '../filter/BeerListFilterBottomSheet';
+import BeerListFilterChipList, {
+  BeerListFilterChipType,
+} from '../beers-filter/BeerListFilterChipList';
+import BeerListFilterBottomSheet from '../beers-filter/BeerListFilterBottomSheet';
 import BeerListSortBottomSheet from '../BeerListSortBottomSheet';
 
 import {
