+++ src/components/beers-filter/BeerListFilterBottomSheet/recoil/atoms/nextBeerListFiler.ts
@@ -1,7 +1,7 @@
 import { atom } from 'recoil';
 
 import { IBeerListFilter } from '@/apis';
-import { BeerListFilterChipType } from '@/components/filter/BeerListFilterChipList/BeerListFilterChipList';
+import { BeerListFilterChipType } from '@/components/beers-filter/BeerListFilterChipList/BeerListFilterChipList';
 
 export const $nextBeerListFilter = atom<IBeerListFilter>({
   key: 'next-beer-list-filter',
