+++ src/components/beers-filter/BeerListFilterFooter/BeerListFilterFooter.tsx
