+++ src/components/beers-filter/BeerListFilterFooter/index.ts
