+++ src/components/beers-filter/BeerTypeFilterItem/BeerTypeFilterItem.stories.tsx
