+++ src/components/beers-filter/BeerTypeFilterItem/BeerTypeFilterItem.tsx
