+++ src/components/beers-filter/BeerTypeFilterItem/index.ts
