+++ src/components/beers-filter/BeerTypeFilterList/BeerTypeFilterList.stories.tsx
