+++ src/components/beers-filter/BeerTypeFilterList/BeerTypeFilterList.tsx
