+++ src/components/beers-filter/BeerTypeFilterList/index.ts
