+++ src/containers/BeerListContainer/recoil/atoms/beerListFilter.ts
@@ -1,7 +1,7 @@
 import { atom } from 'recoil';
 
 import { IBeerListFilter } from '@/apis';
-import { BeerListFilterChipType } from '@/components/filter/BeerListFilterChipList';
+import { BeerListFilterChipType } from '@/components/beers-filter/BeerListFilterChipList';
 
 /** @TODO url 파라미터 값으로 필터 초기값 설정 */
 export const $beerListFilter = atom<IBeerListFilter>({
